protect missing fact for testing + new release

+++ manifests/delete_root_alias.pp
@@ -5,16 +5,19 @@ class system_users::delete_root_alias {
   # user { $user_audit['duplicate']['root_alias']:
   #   ensure => absent,
   # }
+  $users_to_delete = dig($facts, 'user_audit', 'duplicate', 'root_alias')
 
-  $user_audit['duplicate']['root_alias'].each |$remove_user| {
-    # remove from /etc/passwd
-    augeas { "remove root alias from passwd ${remove_user}":
-      changes => "rm /files/etc/passwd/${remove_user}/",
-    }
+  if $users_to_delete {
+    $users_to_delete.each |$remove_user| {
+      # remove from /etc/passwd
+      augeas { "remove root alias from passwd ${remove_user}":
+        changes => "rm /files/etc/passwd/${remove_user}/",
+      }
 
-    # remove from /etc/shadow - will also fire if
-    augeas { "remove root alias from shadow ${remove_user}":
-      changes => "rm /files/etc/shadow/${remove_user}/",
+      # remove from /etc/shadow - will also fire if
+      augeas { "remove root alias from shadow ${remove_user}":
+        changes => "rm /files/etc/shadow/${remove_user}/",
+      }
     }
   }
 }
